+++ src/app/logging/multiLogger.ts
@@ -2,6 +2,7 @@ import { LogLevel } from "./logLevel";
 import { AbstractStringLogger } from "./abstractStringLogger";
 import { ILogger } from "../..";
 import { Guard } from "../utils/guard";
+import { Guid } from "guid-typescript";
 
 /**
  * A logger that is a container for other loggers.
@@ -28,13 +29,14 @@ export class MultiLogger extends AbstractStringLogger {
     /**
      * The real handler for log entries.
      * @param logLevel Entry will be written on this level.
+     * @param guid Guid of the request.
      * @param state The entry to be written. Can be also an object.
      * @param error The error related to this entry.
      * @param formatter Function to create a string message of the state and error.
      */
-    protected logHandler(logLevel: LogLevel, state: string, error: Error, formatter: (s: string, e: Error) => string): void {
+    protected logHandler(logLevel: LogLevel, guid: Guid, state: string, error: Error, formatter: (s: string, guid: Guid, e: Error) => string): void {
         for (const logger of this.loggers) {
-            logger.log(logLevel, state, error, formatter);
+            logger.log(logLevel, guid, state, error, formatter);
         }
     }
 }